SoundRecorder: Update path and extension name before recording.
When the sound recorder does not save and exit, the last sample file
still exists. And the directory and the extension name of the last
sample file is not updated. So the recorder still use the directory
and the extension name of the last sample to create another record.
Before starting another recording task, if the last sample file is not
saved, delete it. Update the storage path and the file extension name
before starting to record another sample.

Change-Id: I9c4da217b470dfe2ada2b2cc40554dc6ab35087b